+++ tools/rpmbuild/SPECS/cuttlefish_user.spec
@@ -10,7 +10,7 @@ URL:            https://github.com/google/android-cuttlefish
 BuildArch:      x86_64
 BuildRoot:      %{_tmppath}/%{name}-%{version}-%{release}-root-%(%{__id_u} -n)
 
-#BuildRequires:  
+BuildRequires:  golang-bin
 Requires:       openssl, cuttlefish-base
 
 %description
@@ -25,8 +25,13 @@ cd ../../../frontend
 ./build-webui.sh
 
 cd ../base/cvd
-bazel query ...
 bazel build cuttlefish:cuttlefish_common --spawn_strategy=local
+
+cd ../../frontend/src/host_orchestrator
+go build
+
+cd ../operator
+go build
 
 %install
 rm -rf $RPM_BUILD_ROOT
@@ -34,29 +39,16 @@ mkdir -p %{buildroot}/etc/default
 mkdir -p %{buildroot}/etc/rc.d/init.d
 mkdir -p %{buildroot}/usr/bin
 mkdir -p %{buildroot}/usr/lib/cuttlefish-common/bin
-mkdir -p %{buildroot}/usr/share/cuttlefish-common/operator/intercept
+mkdir -p %{buildroot}/usr/share/cuttlefish-common/operator/intercept/js
 mkdir -p %{buildroot}/usr/share/cuttlefish-common/operator/static
 mkdir -p %{buildroot}/usr/share/doc/cuttlefish-user
 
-%define srcpath ../../../frontend/src/operator/webui/dist/static
-for filename in $(ls %{srcpath}) ; do
-  install -m 655 %{srcpath}/$filename %{buildroot}/usr/share/cuttlefish-common/operator/static/$filename
-done
 
-%define srcpath ../../../frontend/src/host_orchestrator/host_orchestrator
-for filename in $(ls %{srcpath}) ; do
-  install -m 655 %{srcpath}/$filename %{buildroot}/usr/lib/cuttlefish-common/bin/$filename
-done
-
-%define srcpath ../../../frontend/src/operator/operator
-for filename in $(ls %{srcpath}) ; do
-  install -m 655 %{srcpath}/$filename %{buildroot}/usr/lib/cuttlefish-common/bin/$filename
-done
-
-%define srcpath ../../../frontend/src/operator/intercept
-for filename in $(ls %{srcpath}) ; do
-  install -m 655 %{srcpath}/$filename %{buildroot}/usr/share/cuttlefish-common/operator/$filename
-done
+%define srcpath ../../../frontend/src
+install -m 655 %{srcpath}/host_orchestrator/host_orchestrator %{buildroot}/usr/lib/cuttlefish-common/bin/host_orchestrator
+install -m 655 %{srcpath}/operator/operator %{buildroot}/usr/lib/cuttlefish-common/bin/operator
+install -m 655 %{srcpath}/operator/intercept/js/server_connector.d.ts %{buildroot}/usr/share/cuttlefish-common/operator/intercept/js/server_connector.d.ts
+install -m 655 %{srcpath}//operator/intercept/js/server_connector.js %{buildroot}/usr/share/cuttlefish-common/operator/intercept/js/server_connector.js
 
 %define srcpath ../../../frontend/src/operator/webui/dist/static
 for filename in $(ls %{srcpath}) ; do
@@ -65,9 +57,12 @@ done
 
 
 %files
-/usr/bin/cvd_host_orchestrator
 /usr/lib/cuttlefish-common/bin/host_orchestrator
-/usr/lib/cuttlefish-common/bin/orchestrator
+/usr/lib/cuttlefish-common/bin/operator
+
+/usr/share/cuttlefish-common/operator/intercept/js/server_connector.d.ts
+/usr/share/cuttlefish-common/operator/intercept/js/server_connector.js
+
 /usr/share/cuttlefish-common/operator/static/index.html
 /usr/share/cuttlefish-common/operator/static/3rdpartylicenses.txt
 /usr/share/cuttlefish-common/operator/static/main.*
@@ -82,14 +77,20 @@ done
 #%%doc add-docs-here
 
 %post
-if ! getent passwd _cutf-operator > /dev/null 2>&1 then
-    # The cvdnetwork group is created by cuttlefish-base
-    adduser --system --disabled-password --disabled-login --home /var/empty \
-    --no-create-home --quiet --force-badname --ingroup cvdnetwork _cutf-operator
+ln -sf /usr/lib/cuttlefish-common/bin/host_orchestrator /usr/bin/cvd_host_orchestrator
+# The cvdnetwork group is created by cuttlefish-base
+if ! getent passwd _cutf-operator > /dev/null 2>&1; then
+    adduser --system --home /var/empty --no-create-home --gid cvdnetwork _cutf-operator
 fi
 
+%preun
+rm /usr/bin/cvd_host_orchestrator
+
 %postun
-userdel _cutf-operator
+if getent passwd _cutf-operator > /dev/null 2>&1; then
+    userdel _cutf-operator
+fi
+
 
 %changelog
 * Thu Jul 11 2024 Martin Zeitler <?>
